+++ native_library/android/FirebaseANE/DynamicLinks/src/main/java/com/tuarua/firebase/DynamicLinksANE.java
@@ -21,7 +21,7 @@ import com.adobe.fre.FREContext;
 import com.adobe.fre.FREExtension;
 import com.tuarua.firebase.dynamiclinks.KotlinController;
 
-@SuppressWarnings({"unused", "WeakerAccess"})
+@SuppressWarnings({"unused"})
 public class DynamicLinksANE implements FREExtension {
     private static final String[] FUNCTIONS = {
             "init"