+++ native_library/android/FirebaseANE/VisionText/src/main/java/com/tuarua/firebase/ml/vision/text/KotlinController.kt
@@ -32,7 +32,7 @@ import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.GlobalScope
 import kotlinx.coroutines.launch
 
-@Suppress("unused", "UNUSED_PARAMETER", "UNCHECKED_CAST", "PrivatePropertyName")
+@Suppress("unused", "UNUSED_PARAMETER")
 class KotlinController : FreKotlinMainController {
     private var results: MutableMap<String, FirebaseVisionText> = mutableMapOf()
     private val gson = Gson()