+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ### 0.10.0
+- Restructure API. See example for full changes.
 - AND: Updated to FreKotlin 1.9.5
 - iOS: Updated to FreSwift 4.2.0
 - iOS: Firebase 6.13.0
@@ -7,7 +8,6 @@
 - AND: Crashlytics 2.10.1
 - AND: Vision 24.0.1
 - AND: Firestore 21.3.0
-- AND: Invites 17.0.0
 - AND: Dynamic Links 19.0.0
 - AND: Remote Config 19.0.3
 - AND: Auth 17.0.0
@@ -15,6 +15,7 @@
 - AND: Messaging 20.0.1
 - AND: Play Services 17.1.0
 - AND: Support androidx 1.0.0
+- Remove Invites. Now unsupported by Google
 
 ### 0.6.0
 - Match package structure of official Firebase libraries